requirements.txt
nistitl>=0.6
PyQt5>=5.15.0
Pillow>=10.0.0
wsq>=0.8
pyinstaller>=6.0.0
PyPDF2>=3.0.0
